# openapi.yaml
openapi: 3.1.0
info:
  title: Happy Holiday Icon API
  version: 1.1.0
  description: |
    Creates holiday icons from uploaded images. `POST /api/submit` admits a job
    and returns a bounded presigned S3 POST form. The browser uploads the source
    file directly to S3, then polls the job endpoint while generation and image
    reshaping run asynchronously.
servers:
  - url: http://localhost:3000
    description: Local development
...
paths:
...
  /api/submit:
    post:
      tags: [Jobs]
      operationId: submitImageJob
      summary: Admit an image job and create a bounded source upload form
      description: |
        Authenticates the caller, reserves usage, creates an `UPLOADING` job,
        and returns before the source upload or image generation begins. Submit
        the returned `upload.fields` and original source file as a multipart
        form POST to `upload.url`; S3 enforces the form's content type and byte
        limit. Then poll `GET /api/jobs/{jobId}` for status and final signed
        URLs.

        A valid Cognito access token can be sent as a bearer token or the
        `cognito_access_token` HTTP-only cookie. Requests without a Cognito
        session use the anonymous five-trial allowance; the server returns a
        `session_token` cookie for that path.
      security:
        - bearerAuth: []
        - cognitoAccessToken: []
        - {}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [contentType, contentLength, prompt, style]
              properties:
                contentType:
                  type: string
                  enum: [image/png, image/jpeg, image/webp]
                  description: MIME type of the source file submitted to S3.
                contentLength:
                  type: integer
                  minimum: 1
                  maximum: 10485760
                  description: Source-file byte length. The server pre-validates it against the same configured limit as the S3 upload policy; S3 remains authoritative.
                prompt:
                  type: string
                  minLength: 1
                  maxLength: 500
                  description: A non-whitespace description of the desired icon.
                style:
                  $ref: '#/components/schemas/IconStyle'
      responses:
        '202':
          description: |
            The job is admitted and awaits the source file. The final job state
            will be `READY` or `FAILED` after asynchronous processing.
          headers:
            Set-Cookie:
              description: Sent for anonymous requests to create or refresh the HTTP-only trial session.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubmittedImageJob'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          description: The supplied Cognito access token is invalid or expired.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: The anonymous session has used all five trials.
          headers:
            Set-Cookie:
              description: Refreshes the anonymous HTTP-only trial session cookie.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: A required Cognito, DynamoDB, or S3 dependency is unavailable or not configured.
          headers:
            Set-Cookie:
              description: May be sent for anonymous requests after a trial session is created.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
...
  responses:
    BadRequest:
      description: The request body, form data, or trial session is invalid.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    SubmittedImageJob:
      type: object
      required: [jobId, status, sourceKey, upload, expiresAt]
      properties:
        jobId:
          type: string
          format: uuid
        status:
          type: string
          const: UPLOADING
        sourceKey:
          type: string
          description: Private temporary-bucket object key for the source upload.
          example: uploads/a9f4a6d7-ecf5-448d-a7ce-51954d3a234d/source.png
        upload:
          type: object
          required: [url, fields, maxBytes]
          description: Presigned temporary-bucket POST details, valid for 60 seconds.
          properties:
            url:
              type: string
              format: uri
              description: S3 form action URL.
            fields:
              type: object
              additionalProperties:
                type: string
              description: Fields that must be included unchanged in the S3 multipart form POST.
            maxBytes:
              type: integer
              minimum: 1
              description: Maximum permitted source-file byte size, enforced by S3.
        expiresAt:
          type: integer
          format: int64
          description: Unix epoch seconds when the job expires.
...
    IconStyle:
      type: string
      enum: [playful, minimal, vintage, festive]
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Cognito access token sent in the Authorization header.
    cognitoAccessToken:
      type: apiKey
      in: cookie
      name: cognito_access_token
      description: HTTP-only Cognito access-token cookie established by the login callback.